+++ cmd/dashboards/node/node.go
@@ -1,13 +1,9 @@
-package main
+package node
 
 import (
-	"context"
 	"fmt"
-	"net/http"
-	"os"
 	"strings"
 
-	"github.com/K-Phoen/grabana"
 	"github.com/K-Phoen/grabana/axis"
 	"github.com/K-Phoen/grabana/dashboard"
 	"github.com/K-Phoen/grabana/graph"
@@ -35,8 +31,8 @@ func driveSingleStat(status, triggerColor string, span float32) row.Option {
 	)
 }
 
-func main() {
-	builder := dashboard.New(
+func NewNodeDetailDashboard() dashboard.Builder {
+	return dashboard.New(
 		"Node Detail",
 		common.DashboardTags,
 		common.DashboardAutoRefresh,
@@ -313,19 +309,4 @@ func main() {
 			),
 		),
 	)
-
-	ctx := context.Background()
-	client := grabana.NewClient(&http.Client{}, "http://localhost:3000", grabana.WithBasicAuth("admin", "admin"))
-
-	// create the folder holding the dashboard for the service
-	folder, err := client.FindOrCreateFolder(ctx, "Solidfire")
-	if err != nil {
-		fmt.Printf("Could not find or create folder: %s\n", err)
-		os.Exit(1)
-	}
-
-	if _, err := client.UpsertDashboard(ctx, folder, builder); err != nil {
-		fmt.Printf("Could not create dashboard: %s\n", err)
-		os.Exit(1)
-	}
 }