+++ CHANGELOG.md
@@ -25,8 +25,11 @@ Example Format Below:
 
 ### Added
 - Added CHANGELOG
+- Added new versions of cluster overview, node detail, and volume detail dashboard powered by grabana
 ### Changed
+
 ### Deprecated
+- Marked old json dashboards as deprecated in examples/dashboards/README.md
 ### Removed
 ### Fixed
 ### Security